Automatically sync the configuration on server startup
