+++ crates/cli/src/main.rs
@@ -30,6 +30,7 @@ mod app_state;
 mod commands;
 mod sentry_transport;
 mod server;
+mod sync;
 mod telemetry;
 mod util;
